requirements.txt
# Core DL + CV
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# YOLOv5 torch.hub dependency (prevents pandas import error)
pandas>=2.0.0

# Utilities
tqdm>=4.66.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0.0

# Evaluation + plotting (recommended)
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
scipy>=1.11.0

